V128

+++ app/src/main/java/knf/animeflv/Interfaces/WaitDownloadCallback.java
@@ -6,7 +6,7 @@ import java.util.List;
  * Created by Jordy on 01/04/2016.
  */
 public interface WaitDownloadCallback {
-    void onAllCapsDownload(String aid, List<Integer> list);
+    void onAllCapsDownload(String aid, List<Float> list);
 
-    void onSingleCapDownload(String aid, int num);
+    void onSingleCapDownload(String aid, float num);
 }
